move func to docker package
