+++ internal/docker/docker_test.go
@@ -17,6 +17,8 @@ package docker
 import (
 	"context"
 	"fmt"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/googleapis/librarian/internal/config"
@@ -66,6 +68,7 @@ func TestDockerRun(t *testing.T) {
 		testRepoRoot        = "testRepoRoot"
 	)
 
+	state := &config.PipelineState{}
 	cfg := &config.Config{}
 	cfgInDocker := &config.Config{
 		HostMount: "hostDir:localDir",
@@ -82,7 +85,7 @@ func TestDockerRun(t *testing.T) {
 				Image: testImage,
 			},
 			runCommand: func(ctx context.Context, d *Docker) error {
-				return d.Generate(ctx, cfg, testAPIRoot, testOutput, testGenerateRequest, testGeneratorInput, testLibraryID)
+				return d.Generate(ctx, cfg, state, testAPIRoot, testOutput, testGenerateRequest, testGeneratorInput, testLibraryID)
 			},
 			want: []string{
 				"run", "--rm",
@@ -105,7 +108,7 @@ func TestDockerRun(t *testing.T) {
 				Image: testImage,
 			},
 			runCommand: func(ctx context.Context, d *Docker) error {
-				return d.Generate(ctx, cfgInDocker, testAPIRoot, "hostDir", testGenerateRequest, testGeneratorInput, testLibraryID)
+				return d.Generate(ctx, cfgInDocker, state, testAPIRoot, "hostDir", testGenerateRequest, testGeneratorInput, testLibraryID)
 			},
 			want: []string{
 				"run", "--rm",
@@ -173,4 +176,91 @@ func TestDockerRun(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestToGenerateRequestJSON(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		name      string
+		state     *config.PipelineState
+		expectErr bool
+	}{
+		{
+			name: "successful-marshaling-and-writing",
+			state: &config.PipelineState{
+				ImageTag: "v1.0.0",
+				Libraries: []*config.LibraryState{
+					{
+						ID:                        "google-cloud-go",
+						CurrentVersion:            "1.0.0",
+						GenerationAutomationLevel: config.AutomationLevelAutomatic,
+						APIPaths:                  []string{"google/cloud/compute/v1"},
+					},
+					{
+						ID:                        "google-cloud-storage",
+						CurrentVersion:            "1.2.3",
+						GenerationAutomationLevel: config.AutomationLevelManualReview,
+						APIPaths:                  []string{"google/storage/v1"},
+					},
+				},
+				IgnoredAPIPaths: []string{"google/cloud/ignored/v1"},
+			},
+			expectErr: false,
+		},
+		{
+			name:      "empty-pipelineState",
+			state:     &config.PipelineState{},
+			expectErr: false,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			tempDir := t.TempDir()
+			filePath := filepath.Join(tempDir, "generate-request.json")
+
+			err := toGenerateRequestJSON(tc.state, filePath)
+
+			if tc.expectErr {
+				if err == nil {
+					t.Errorf("toGenerateRequestJSON() expected an error but got nil")
+				}
+				return // Test case expects error, so no further checks
+			}
+
+			if err != nil {
+				t.Fatalf("toGenerateRequestJSON() unexpected error: %v", err)
+			}
+
+			// Verify the file content
+			gotBytes, err := os.ReadFile(filePath)
+			if err != nil {
+				t.Fatalf("Failed to read generated file: %v", err)
+			}
+
+			fileName := fmt.Sprintf("%s.json", tc.name)
+			wantBytes, readErr := os.ReadFile(filepath.Join("..", "..", "testdata", fileName))
+			if readErr != nil {
+				t.Fatalf("Failed to read expected state for comparison: %v", readErr)
+			}
+
+			if diff := cmp.Diff(string(wantBytes), string(gotBytes)); diff != "" {
+				t.Errorf("Generated JSON mismatch (-want +got):\n%s", diff)
+			}
+		})
+	}
+
+	t.Run("Error creating file (invalid path)", func(t *testing.T) {
+		t.Parallel()
+		// Attempt to write to a path where we don't have permissions or a non-existent dir
+		invalidPath := filepath.Join("/nonexistent_dir_for_test", "generate-request.json")
+		state := &config.PipelineState{ImageTag: "test"}
+		err := toGenerateRequestJSON(state, invalidPath)
+		if err == nil {
+			t.Error("toGenerateRequestJSON() expected an error for invalid path, got nil")
+		}
+	})
 }
